+++ lightstep/sdk/metric/aggregator/histogram/structure/exponential.go
@@ -12,14 +12,55 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package histogram // import "go.opentelemetry.io/otel/sdk/metric/aggregator/histogram"
+package structure // import "github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator/histogram/structure"
 
 import (
 	"fmt"
 	"math/bits"
+	"sync"
+
+	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator/aggregation"
+	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/number"
+	"go.opentelemetry.io/otel/sdk/metric/aggregator/exponential/mapping"
+	"go.opentelemetry.io/otel/sdk/metric/aggregator/exponential/mapping/exponent"
+	"go.opentelemetry.io/otel/sdk/metric/aggregator/exponential/mapping/logarithm"
 )
 
 type (
+	// State observes counts observations in exponentially-spaced
+	// buckets.  It is configured with a maximum scale factor
+	// which determines resolution.  Scale is automatically
+	// adjusted to accommodate the range of input data.
+	State[N int64 | float64] struct {
+		lock    sync.Mutex
+		maxSize int32
+		data[N]
+	}
+
+	data[N int64 | float64] struct {
+		// sum is the sum of all Updates reflected in the
+		// aggregator.  It has the same type number as the
+		// corresponding sdkinstrument.Descriptor.
+		sum N
+		// count is incremented by 1 per Update.
+		count uint64
+		// zeroCount is incremented by 1 when the measured
+		// value is exactly 0.
+		zeroCount uint64
+		// min is set when count > 0
+		min N
+		// max is set when count > 0
+		max N
+		// positive holds the positive values
+		positive buckets
+		// negative holds the negative values in these buckets
+		// by their absolute value.
+		negative buckets
+		// mapping corresponds to the current scale, is shared
+		// by both positive and negative ranges.
+		mapping mapping.Mapping
+	}
+
 	// buckets stores counts for measurement values in the range
 	// (0, +Inf).
 	buckets struct {
@@ -40,9 +81,9 @@ type (
 		indexEnd int32
 	}
 
-	// valueType is an interface constraint for the numeric type
+	// ValueType is an interface constraint for the numeric type
 	// aggregated by this histogram.
-	valueType interface {
+	ValueType interface {
 		int64 | float64
 	}
 
@@ -87,7 +128,144 @@ type (
 		low  int32
 		high int32
 	}
+
+	Int64   = State[int64]
+	Float64 = State[float64]
 )
+
+func NewFloat64(cfg Config, values ...float64) *Float64 {
+	return newHist[float64](cfg, values)
+}
+
+func NewInt64(cfg Config, values ...int64) *Int64 {
+	return newHist[int64](cfg, values)
+}
+
+func newHist[N number.Any](cfg Config, values []N) *State[N] {
+	state := &State[N]{}
+
+	state.Init(cfg)
+
+	for _, val := range values {
+		state.Update(val)
+	}
+	return state
+}
+
+func (h *State[N]) Init(cfg Config) {
+	h.maxSize = cfg.maxSize
+
+	mapping, _ := newMapping(logarithm.MaxScale)
+	h.mapping = mapping
+}
+
+// Sum implements aggregation.Histogram.
+func (h *State[N]) Sum() N {
+	return h.sum
+}
+
+// Min implements aggregation.Histogram.
+func (h *State[N]) Min() N {
+	return h.min
+}
+
+// Max implements aggregation.Histogram.
+func (h *State[N]) Max() N {
+	return h.max
+}
+
+// Count implements aggregation.Histogram.
+func (h *State[N]) Count() uint64 {
+	return h.count
+}
+
+// Scale implements aggregation.Histogram.
+func (h *State[N]) Scale() int32 {
+	if h.data.count == h.data.zeroCount {
+		// all zeros! scale doesn't matter, use zero.
+		return 0
+	}
+	return h.data.mapping.Scale()
+}
+
+// ZeroCount implements aggregation.Histogram.
+func (h *State[N]) ZeroCount() uint64 {
+	return h.data.zeroCount
+}
+
+// Positive implements aggregation.Histogram.
+func (h *State[N]) Positive() *buckets {
+	return &h.data.positive
+}
+
+// Negative implements aggregation.Histogram.
+func (h *State[N]) Negative() *buckets {
+	return &h.data.negative
+}
+
+// Offset implements aggregation.Bucket.
+func (b *buckets) Offset() int32 {
+	return b.indexStart
+}
+
+// Len implements aggregation.Bucket.
+func (b *buckets) Len() uint32 {
+	if b.backing == nil {
+		return 0
+	}
+	if b.indexEnd == b.indexStart && b.At(0) == 0 {
+		return 0
+	}
+	return uint32(b.indexEnd - b.indexStart + 1)
+}
+
+// At returns the count of the bucket at a position in the logical
+// array of counts.
+func (b *buckets) At(pos0 uint32) uint64 {
+	pos := pos0
+	bias := uint32(b.indexBase - b.indexStart)
+
+	if pos < bias {
+		pos += uint32(b.backing.size())
+	}
+	pos -= bias
+
+	return b.backing.countAt(pos)
+}
+
+// clearState resets a histogram to the empty state without changing
+// backing array.  Scale is reset if there are no range limits.
+func (h *State[N]) clearState() {
+	h.positive.clearState()
+	h.negative.clearState()
+	h.sum = 0
+	h.count = 0
+	h.zeroCount = 0
+	h.min = 0
+	h.max = 0
+	h.mapping, _ = newMapping(logarithm.MaxScale)
+}
+
+// clearState zeros the backing array.
+func (b *buckets) clearState() {
+	b.indexStart = 0
+	b.indexEnd = 0
+	b.indexBase = 0
+	if b.backing != nil {
+		b.backing.reset()
+	}
+}
+
+func newMapping(scale int32) (mapping.Mapping, error) {
+	if scale <= 0 {
+		return exponent.NewMapping(scale)
+	}
+	return logarithm.NewMapping(scale)
+}
+
+func (h *State[N]) Kind() aggregation.Kind {
+	return aggregation.HistogramKind
+}
 
 // Move atomically copies and resets `s`.  The modification to `dest`
 // is not synchronized.
@@ -114,7 +292,18 @@ func (s *State[N]) Move(dest *State[N]) {
 	}
 }
 
-// Update adds the recorded measurement to the current data set.
+// Copy atomically copies `s`.  The modification to `dest` is not
+// synchronized.
+func (s *State[N]) Copy(dest *State[N]) {
+	s.lock.Lock()
+	defer s.lock.Unlock()
+
+	dest.clearState()
+	dest.Merge(s)
+}
+
+// UpdateByIncr supports updating a histogram with a non-negative
+// increment.
 func (s *State[N]) Update(number N) {
 	s.UpdateByIncr(number, 1)
 }
